all: Add "user-space" apiserver checkpointing
This patch adds a new binary: "checkpoint". The checkpoint program will
ensure that the latest apiserver manifest is checkpointed, in case of a
system / apiserver crash.

It is implemented via static manifests. The checkpoint program will
store a pod manifest on disk. When it detects an apiserver is not
running, it will move that file into the directory that was specified as
the kubelet's config dir. From there, the kubelet will see that pod
manifest and run it. Once the program detects both our temporary
apiserver and self-hosted apiserver is running, it will remove the
manifest from the config dir, causing the kubelet to kill it and allow
the self-hosted apiserver to take over.