all: Include checkpointer
This patch simply includes the checkpointer, and bumps the hyperkube
image version to the latest beta release, which carries a patch needed
for checkpointing to work.

Checkpointing is enabled automatically. The API server pod now contains
a "checkpoint-installer" container, which will install the checkpoint
static manifest, at which point the kubelet will run the checkpoint pod.