cmd/checkpoint: Verify API server running
Adds a check to ensure the API is actually running, responding to
requests, when the checkpoint service notices the pod manifest. Came
accross an issue when upgrading a cluster where the API server daemonset
was removed, and the API server brought down, but the kubelet still
returned that pod from its read-only API.

Now, if we notice that an API server pod is returned from kubelet (the
actual, self-hosted one, not a temp one) then we also check that it is
up and accepting requests, otherwise we assume it to be dead.