update release guides
Even though it makes sense to only bump the quickstart guides after we
have cut a bootkube release, the conformance scripts use the quickstart
guide by overwriting the bootkube container image to use but not the
kubelet version. Lets make this change and accept that the quickstart
guides will have a temporary version skew between updating hyperkube and
releasing the image, but the the person doing the release via the
conformance scripts will at least run the correct host hyperkube.