hack/quickstart: Add automatic self-hosted multi-master loadbalancer
If MULTI_MASTER=true is specified, init-{master,node}.sh add a static
nginx-proxy manifest, change kubeconfig server to 127.0.0.1:6443 and
add a nginx.conf file.

For the first master (init-master.sh) it also inject a manifest file
which add nginx-conf-syncer and nginx-conf-checkpointer.
The nginx.conf for the first node just redirect 127.0.0.1:6443 -> 127.0.0.1:443.

nginx-conf-syncer generate a nginx.conf every minute, by using the
template "nginx.conf.template" with sigil from the configmap nginx.conf.
The nginx.conf is saved in the same configmap as "nginx.conf".
It get the master node by looking at the node with the label:
node-role.kubernetes.io/master

nginx-conf-syncer pull the "nginx.conf" and compare it to the
nginx.conf on disk, if they are different it update the file on
disk and tell nginx to reload the config.

Could fix: #311