hack/quickstart: Make it easy to use COREOS_PUBLIC_IPV4 as hostname-override
With PUBLIC_IP=true set, --hostname-override will be set to
${COREOS_PUBLIC_IPV4} instead of ${COREOS_PRIVATE_IPV4}.
This make most sense when using the quickstart scripts with a
cloud provider like DigitalOcean and you don't want to limit yourself
to a single location.
Maybe a bit out of scope for the quickstart script, but they are
very simple and get the job done.