Makefile: Use 'go env GOOS' instead of uname
And change the variable from LOCAL_OS to GOOS.

We've been using uname for this since LOCAL_OS landed in 04848d13 (Add
release build process, 2016-05-27, #47), but there's no need to reach
outside of Go for this information.  Using 'go env' also ensures we
get the canonical phrasing without the tr pipe.

I'm using ?= (documented in [1]) to avoid an expensive shell operation
when we don't actually need the result (because a different target was
called or because the user provided their own value).  That doesn't
change anything (yet), because we're still consuming this variable in
the run-% prereqs, and Make expands prerequisites immediately when
loading a Makefile [2], but now we're a step closer to only evaluating
this for targets that need it.

The uname/tr -> 'go env' change affects the run-% prereqs if someone
calls make with GOOS set to a non-native system.  But I don't think
that use case is likely, so I've added a guard (with the old uname/tr
logic) to fail it cleanly instead of trying to guess what was
intended.

[1]: https://www.gnu.org/software/make/manual/html_node/Setting.html
[2]: https://www.gnu.org/software/make/manual/html_node/Reading-Makefiles.html#Rule-Definition