pkg/asset/k8s: Evenly weight chars in newBootstrapToken
The 256 % len(charset) remainder unevenly biased characters listed
earlier in the charset array.  For example, 256 is 7 * 36 + 4, so a
36-character charset that naively used int(b) % len(charset) would
include the first four characters (0 through 3) 14% (8/7) more
frequently than the other characters.

The new integer approach is unbiased and a more efficient use of
entropy, because we no longer collapse 256 possible values into 36
possible values.  With the current 36-character token charset, this
reduces our read from 22 bytes to 15 bytes.  The size calculation for
that 15 bytes is from:

  256^sizef = len(charset)^len(token)
  log2(256^sizef) > log2(len(charset)^len(token))       # log2 both sides
  sizef * log2(256) > len(token) * log2(len(charset))   # extract exponents
  sizef > len(token) * log2(len(charset)) / log2(256)   # divide by log2(256)
  size = ceil(len(token)*log2(len(charset))/log2(256))  # must read full bytes

for our 36-character token charset and 6+16 token length, that comes
to:

  size = ceil((6+16)*log2(36)/log2(256)) = 15

SetBytes converts the full input slice into a big integer.  There may
be a more efficient approach to this that consumes the input
incrementally, but SetBytes is convenient.  We can always polish more
later if we want a more efficient approach.

The endian-ness doesn't matter for our current use case, because the
input data is random.  I've documented it to make independently
calculating the expected test values easier, but we can flip the
endian-ness of the internal helper later on if it's convenient to do
so.