pkg/poll: Add a Poll helper that respects Context
This is like the k8s.io/apimachinery/pkg/util Poll, except it takes a
Context argument to allow canceling for any reason.  The k8s utility
only cancels the poller on timeouts.

The ConditionFunc type and godoc are from the k8s package (which is
also licensed under the Apache 2.0), although I've added the Context
argument locally in case slow condition functions want to watch for
finishing contexts.

The Poll godoc is heavily based on the k8s Poll godoc.