e2e: Transition from k8s wait package to our poll package
The functional differences between the packages matter less for tests
than they do for the bootkube package and executable.  But now that
the non-test consumers are using our package, transitioning the tests
will let us drop a direct external dependency.  In this case we still
depend on the k8s package via k8s.io/client-go/tools/cache, so we
can't remove the vendored wait package.  But we can bump our vendored
packages without worrying about upstream changes to the k8s wait
package breaking our usage.

We can also drop PollTimeoutForNginx, because its consumers are all
being called with contexts that time out in less than 10 minutes:

* NewNginx
  * In TestNetwork, NewNginx is being called as a Poll condition, and
    the entire Poll times out after networkPingPollTimeout (5
    minutes).
	* In TestSmoke, I've set the whole test to time out after 10
    minutes.
* IsReachable is always called as part of a Poll condition, and I've
  set it up to timeout before the poll interval completes (either one
  second before or 200 ms before depending on the location).

* Delete
  * In NewNginx, it inherits the context that the caller passed in.
    The NewNginx timeouts are already under 10 minutes, as discussed
    above.
  * In TestNetwork, I've set a new timeout for the deferred Delete
    call using networkPingPollTimeout.
  * In TestSmoke, it's getting the whole-test 10 minute timeout.

This is a bit of a shift for the Deletes called from NewNginx and
TestSmoke , because they might be called with an already-timed-out
Context.  In that case, the delete will only succeed if the initial
(immediate) condition call executes successfully.  If that causes
problems, we can adjust those executions to extend the deadline,
followwing the approach I'm using in TestNetwork.