AWS Quickstart: mention TAG_MASTER
There should be a mention of adding additional masters. I understand this is a difficult thing to place correctly, since it's important to note that boot-kube is only necessary for the first node in the cluster.  This is why I mentioned it last.  It could also be a comment line in the **Add Workers** section if you think it's too confusing down here.